move git_context out
